# Saving memory in store_poes