By storing one IMT at the time. This reduces by 7 times the required memory in the master in the Global Hazard Model, with an insignificant penalty on `store_poes` and an actually faster `postclassical`:
```
# before
| calc_53976, maxmem=142.3 GB | time_sec | memory_mb | counts    |
|-----------------------------+----------+-----------+-----------|
| total classical             | 29_649   | 103.9     | 364       |
| get_poes                    | 16_404   | 0.0       | 1_625_769 |
| composing pnes              | 6_300    | 0.0       | 1_625_769 |
| computing mean_std          | 5_903    | 0.0       | 11_155    |
| total postclassical         | 572.3    | 1_054     | 29        |
| ClassicalCalculator.run     | 545.3    | 7_156     | 1         |
| read PoEs                   | 98.4     | 1_028     | 29        |
| storing PoEs                | 87.1     | 2_543     | 1         |

# after
| calc_53978, maxmem=142.3 GB | time_sec | memory_mb | counts    |
|-----------------------------+----------+-----------+-----------|
| total classical             | 29_748   | 105.4     | 364       |
| get_poes                    | 16_423   | 0.0       | 1_625_769 |
| composing pnes              | 6_358    | 0.0       | 1_625_769 |
| computing mean_std          | 5_929    | 0.0       | 11_155    |
| total postclassical         | 559.8    | 1_186     | 29        |
| ClassicalCalculator.run     | 558.8    | 4_612     | 1         |
| read PoEs                   | 94.6     | 1_186     | 29        |
| storing PoEs                | 93.6     | 0.07422   | 1         |
```